requirements.txt
aioice==0.7.5
aiortc==1.2.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
av==8.0.3
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
crc32c==2.2
cryptography==3.4.7
decorator==5.0.7
defusedxml==0.7.1
dnspython==2.1.0
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.0
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.9
notebook==6.3.0
numpy==1.20.3
opencv-python==4.5.2.52
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.16.0
pyarrow==4.0.0
pycparser==2.20
pydeck==0.6.2
pyee==8.1.0
Pygments==2.9.0
pylibsrtp==0.6.8
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.25.1
Send2Trash==1.5.0
six==1.16.0
smmap==4.0.0
streamlit==0.81.1
streamlit-webrtc==0.20.0
terminado==0.9.5
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
watchdog==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
